backend/requirements.txt
# Web Framework
Flask>=3.0.0
flask-cors>=4.0.0
flasgger>=0.9.5

# Sentinel Hub
sentinelhub>=3.10.0

# Image Processing
numpy>=2.3.0
opencv-python>=4.9.0.80
Pillow>=10.0.0
scipy>=1.11.0

# Geospatial
# rasterio>=1.3.0  # Requiere GDAL - instalar manualmente si es necesario
shapely>=2.0.0
pyproj>=3.6.0
geojson>=3.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development
python-dateutil>=2.8.0
